fix: Update tag property to tags in PaintingDetails component
