+++ src/pages/Details/PaintingDetail/index.tsx
@@ -32,7 +32,7 @@ const defaultPainting: Painting = {
     images: [],
   },
   state: "",
-  tag: [],
+  tags: [],
 };
 
 const PaintingDetails = () => {
@@ -178,7 +178,7 @@ const PaintingDetails = () => {
           <div className="topic-wrapper">
             <h2 className="tags-title">Tags</h2>
             <div className="tags-wrapper">
-              <Tags tags={data.tag} />
+              <Tags tags={data.tags} />
             </div>
           </div>
         </div>
