+++ src/pages/RegisterPage/index.tsx
@@ -5,11 +5,13 @@ import { Container, Content } from "./styles";
 import googleLogo from "../../assets/google_symbol.svg.png";
 import axios from "axios";
 import toast, { Toaster } from "react-hot-toast";
+import { useNavigate } from "react-router-dom";
 
 const RegisterPage = () => {
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const navigate = useNavigate();
 
   const handleRegister = () => {
     axios
@@ -17,6 +19,13 @@ const RegisterPage = () => {
         name,
         email,
         password,
+      })
+      .then((response) => {
+        console.log("User registered:", response.data);
+        toast.success("Usuário registrado com sucesso!");
+        setTimeout(() => {
+          navigate("/login");
+        }, 2000);
       })
       .catch((error) => {
         console.error("Error registering user:", error);
